[pipeline](rpc) support closure reuse in pipeline exec engine
